@@ -34,7 +34,7 @@ namespace Microsoft.Azure.Devices.E2ETests.Provisioning
         [TestCategory("Proxy")]
         public async Task ProvisioningServiceClient_IndividualEnrollments_Query_HttpWithProxy_Ok()
         {
-            await ProvisioningServiceClientE2ETests.ProvisioningServiceClient_IndividualEnrollments_Query_Ok(s_proxyServerAddress).ConfigureAwait(false);
+            await ProvisioningServiceClient_IndividualEnrollments_Query_Ok(s_proxyServerAddress).ConfigureAwait(false);
         }
 
         [LoggedTestMethod]
@@ -196,8 +196,12 @@ namespace Microsoft.Azure.Devices.E2ETests.Provisioning
                 attestationMechanism.Type.Should().Be(AttestationMechanismType.X509);
 
                 var x509Attestation = (X509Attestation)attestationMechanism.GetAttestation();
-                x509Attestation.GetPrimaryX509CertificateInfo().SHA1Thumbprint.Should().Be(((X509Attestation)enrollmentGroup.Attestation).GetPrimaryX509CertificateInfo().SHA1Thumbprint);
-                x509Attestation.GetSecondaryX509CertificateInfo().SHA1Thumbprint.Should().Be(((X509Attestation)enrollmentGroup.Attestation).GetSecondaryX509CertificateInfo().SHA1Thumbprint);
+                x509Attestation.GetPrimaryX509CertificateInfo().SHA1Thumbprint
+                    .Should()
+                    .Be(((X509Attestation)enrollmentGroup.Attestation).GetPrimaryX509CertificateInfo().SHA1Thumbprint);
+                x509Attestation.GetSecondaryX509CertificateInfo().SHA1Thumbprint
+                    .Should()
+                    .Be(((X509Attestation)enrollmentGroup.Attestation).GetSecondaryX509CertificateInfo().SHA1Thumbprint);
             }
         }
 
@@ -220,7 +224,14 @@ namespace Microsoft.Azure.Devices.E2ETests.Provisioning
 
         public async Task ProvisioningServiceClient_IndividualEnrollments_Create_Ok(string proxyServerAddress, AttestationMechanismType attestationType)
         {
-            await ProvisioningServiceClient_IndividualEnrollments_Create_Ok(proxyServerAddress, attestationType, null, AllocationPolicy.Hashed, null, null).ConfigureAwait(false);
+            await ProvisioningServiceClient_IndividualEnrollments_Create_Ok(
+                    proxyServerAddress,
+                    attestationType,
+                    null,
+                    AllocationPolicy.Hashed,
+                    null,
+                    null)
+                .ConfigureAwait(false);
         }
 
         public async Task ProvisioningServiceClient_IndividualEnrollments_Create_Ok(
@@ -229,23 +240,26 @@ namespace Microsoft.Azure.Devices.E2ETests.Provisioning
             ReprovisionPolicy reprovisionPolicy,
             AllocationPolicy allocationPolicy,
             CustomAllocationDefinition customAllocationDefinition,
-            ICollection<string> iotHubsToProvisionTo)
+            IList<string> iotHubsToProvisionTo)
         {
             using ProvisioningServiceClient provisioningServiceClient = CreateProvisioningService(proxyServerAddress);
             string registrationId = AttestationTypeToString(attestationType) + "-" + Guid.NewGuid();
 
             IndividualEnrollment individualEnrollment = await CreateIndividualEnrollmentAsync(
-                provisioningServiceClient,
-                registrationId,
-                attestationType,
-                null,
-                reprovisionPolicy,
-                allocationPolicy,
-                customAllocationDefinition,
-                iotHubsToProvisionTo,
-                null,
-                Logger).ConfigureAwait(false);
-            IndividualEnrollment individualEnrollmentResult = await provisioningServiceClient.GetIndividualEnrollmentAsync(individualEnrollment.RegistrationId).ConfigureAwait(false);
+                    provisioningServiceClient,
+                    registrationId,
+                    attestationType,
+                    null,
+                    reprovisionPolicy,
+                    allocationPolicy,
+                    customAllocationDefinition,
+                    iotHubsToProvisionTo,
+                    null,
+                    Logger)
+                .ConfigureAwait(false);
+            IndividualEnrollment individualEnrollmentResult = await provisioningServiceClient
+                .GetIndividualEnrollmentAsync(individualEnrollment.RegistrationId)
+                .ConfigureAwait(false);
             Assert.AreEqual(individualEnrollmentResult.ProvisioningStatus, ProvisioningStatus.Enabled);
 
             if (reprovisionPolicy != null)
@@ -277,7 +291,7 @@ namespace Microsoft.Azure.Devices.E2ETests.Provisioning
             ReprovisionPolicy reprovisionPolicy,
             AllocationPolicy allocationPolicy,
             CustomAllocationDefinition customAllocationDefinition,
-            ICollection<string> iothubs)
+            IList<string> iothubs)
         {
             string groupId = s_devicePrefix + AttestationTypeToString(attestationType) + "-" + Guid.NewGuid();
             using ProvisioningServiceClient provisioningServiceClient = CreateProvisioningService(proxyServerAddress);
@@ -327,13 +341,14 @@ namespace Microsoft.Azure.Devices.E2ETests.Provisioning
             ReprovisionPolicy reprovisionPolicy,
             AllocationPolicy allocationPolicy,
             CustomAllocationDefinition customAllocationDefinition,
-            ICollection<string> iotHubsToProvisionTo,
+            IList<string> iotHubsToProvisionTo,
             Devices.Provisioning.Service.DeviceCapabilities capabilities,
             MsTestLogger logger)
         {
             Attestation attestation;
             IndividualEnrollment individualEnrollment;
             IndividualEnrollment createdEnrollment = null;
+
             switch (attestationType)
             {
                 case AttestationMechanismType.Tpm:
@@ -346,7 +361,7 @@ namespace Microsoft.Azure.Devices.E2ETests.Provisioning
                             AllocationPolicy = allocationPolicy,
                             ReprovisionPolicy = reprovisionPolicy,
                             CustomAllocationDefinition = customAllocationDefinition,
-                            IotHubs = iotHubsToProvisionTo
+                            IotHubs = iotHubsToProvisionTo,
                         };
 
                         IndividualEnrollment temporaryCreatedEnrollment = null;
@@ -354,7 +369,9 @@ namespace Microsoft.Azure.Devices.E2ETests.Provisioning
                             .RetryOperationsAsync(
                                 async () =>
                                 {
-                                    temporaryCreatedEnrollment = await provisioningServiceClient.CreateOrUpdateIndividualEnrollmentAsync(individualEnrollment).ConfigureAwait(false);
+                                    temporaryCreatedEnrollment = await provisioningServiceClient
+                                        .CreateOrUpdateIndividualEnrollmentAsync(individualEnrollment)
+                                        .ConfigureAwait(false);
                                 },
                                 s_provisioningServiceRetryPolicy,
                                 s_retryableExceptions,
@@ -363,7 +380,7 @@ namespace Microsoft.Azure.Devices.E2ETests.Provisioning
 
                         if (temporaryCreatedEnrollment == null)
                         {
-                            throw new ArgumentException($"The enrollment entry with registration Id {registrationId} could not be created, exiting test.");
+                            throw new ArgumentException($"The enrollment entry with registration Id {registrationId} could not be created; exiting test.");
                         }
 
                         attestation = new TpmAttestation(base64Ek);
@@ -373,7 +390,9 @@ namespace Microsoft.Azure.Devices.E2ETests.Provisioning
                             .RetryOperationsAsync(
                                 async () =>
                                 {
-                                    createdEnrollment = await provisioningServiceClient.CreateOrUpdateIndividualEnrollmentAsync(temporaryCreatedEnrollment).ConfigureAwait(false);
+                                    createdEnrollment = await provisioningServiceClient
+                                        .CreateOrUpdateIndividualEnrollmentAsync(temporaryCreatedEnrollment)
+                                        .ConfigureAwait(false);
                                 },
                                 s_provisioningServiceRetryPolicy,
                                 s_retryableExceptions,
@@ -382,7 +401,7 @@ namespace Microsoft.Azure.Devices.E2ETests.Provisioning
 
                         if (createdEnrollment == null)
                         {
-                            throw new ArgumentException($"The enrollment entry with registration Id {registrationId} could not be updated, exiting test.");
+                            throw new ArgumentException($"The enrollment entry with registration Id {registrationId} could not be updated; exiting test.");
                         }
 
                         return createdEnrollment;
@@ -415,7 +434,9 @@ namespace Microsoft.Azure.Devices.E2ETests.Provisioning
                 .RetryOperationsAsync(
                     async () =>
                     {
-                        createdEnrollment = await provisioningServiceClient.CreateOrUpdateIndividualEnrollmentAsync(individualEnrollment).ConfigureAwait(false);
+                        createdEnrollment = await provisioningServiceClient
+                            .CreateOrUpdateIndividualEnrollmentAsync(individualEnrollment)
+                            .ConfigureAwait(false);
                     },
                     s_provisioningServiceRetryPolicy,
                     s_retryableExceptions,
@@ -424,7 +445,7 @@ namespace Microsoft.Azure.Devices.E2ETests.Provisioning
 
             if (createdEnrollment == null)
             {
-                throw new ArgumentException($"The enrollment entry with registration Id {registrationId} could not be created, exiting test.");
+                throw new ArgumentException($"The enrollment entry with registration Id {registrationId} could not be created; exiting test.");
             }
 
             return createdEnrollment;
@@ -437,11 +458,12 @@ namespace Microsoft.Azure.Devices.E2ETests.Provisioning
             ReprovisionPolicy reprovisionPolicy,
             AllocationPolicy allocationPolicy,
             CustomAllocationDefinition customAllocationDefinition,
-            ICollection<string> iothubs,
+            IList<string> iothubs,
             Devices.Provisioning.Service.DeviceCapabilities capabilities,
             MsTestLogger logger)
         {
             Attestation attestation;
+
             switch (attestationType)
             {
                 case AttestationMechanismType.Tpm:
