@@ -2,7 +2,8 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System.Collections.Generic;
-using System.Diagnostics.CodeAnalysis;
+using System.ComponentModel;
+using System.Linq;
 using Newtonsoft.Json;
 
 namespace Microsoft.Azure.Devices.Serialization
@@ -13,22 +14,40 @@ namespace Microsoft.Azure.Devices.Serialization
     public class BasicDigitalTwin
     {
         /// <summary>
-        /// The unique Id of the digital twin. This is present at the root of every digital twin.
+        /// The unique Id of the digital twin.
         /// </summary>
+        /// <remarks>
+        /// This is present at the root of every digital twin.
+        /// </remarks>
         [JsonProperty("$dtId")]
         public string Id { get; set; }
 
         /// <summary>
-        /// Information about the model a digital twin conforms to. This field is present on every digital twin.
+        /// Information about the model a digital twin conforms to.
         /// </summary>
+        /// <remarks>
+        /// This field is present on every digital twin.
+        /// </remarks>
         [JsonProperty("$metadata")]
-        public DigitalTwinMetadata Metadata { get; set; } = new DigitalTwinMetadata();
+        public DigitalTwinMetadata Metadata { get; set; }
 
         /// <summary>
-        /// Additional properties of the digital twin. This field will contain any properties of the digital twin that are not already defined by the other strong types of this class.
+        /// Additional properties of the digital twin.
         /// </summary>
+        /// <remarks>
+        /// This field will contain any properties of the digital twin that are not already defined by the other strong types of this class.
+        /// </remarks>
         [JsonExtensionData]
-        [SuppressMessage("Usage", "CA2227:Collection properties should be read only", Justification = "Public facing property cannot be modified since it will be a breaking change.")]
         public IDictionary<string, object> CustomProperties { get; set; } = new Dictionary<string, object>();
+
+        /// <summary>
+        /// For use in serialization.
+        /// </summary>
+        /// <seealso href="https://www.newtonsoft.com/json/help/html/ConditionalProperties.htm#ShouldSerialize"/>
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        public bool ShouldSerializeCustomProperties()
+        {
+            return CustomProperties != null && CustomProperties.Any();
+        }
     }
 }
